Revert "[web] Switch to doubles (#37336)" (flutter/engine#37609)
This reverts commit e3bed0a86df37652c27e3667fcc99bfac75a741e.

+++ engine/src/flutter/lib/web_ui/lib/src/engine/canvaskit/image.dart
@@ -111,7 +111,7 @@ Future<Uint8List> fetchImage(
   if (chunkCallback != null) {
     request.addEventListener('progress', allowInterop((DomEvent event)  {
       event = event as DomProgressEvent;
-      chunkCallback.call(event.loaded!.toInt(), event.total!.toInt());
+      chunkCallback.call(event.loaded!, event.total!);
     }));
   }
 
@@ -123,7 +123,7 @@ Future<Uint8List> fetchImage(
   }));
 
   request.addEventListener('load', allowInterop((DomEvent event) {
-    final int status = request.status!.toInt();
+    final int status = request.status!;
     final bool accepted = status >= 200 && status < 300;
     final bool fileUri = status == 0; // file:// URIs have status of 0.
     final bool notModified = status == 304;
@@ -173,8 +173,8 @@ class CkImage implements ui.Image, StackTraceDebugger {
             'be able to resurrect it once it has been garbage collected.');
         return;
       }
-      final int originalWidth = skImage.width().toInt();
-      final int originalHeight = skImage.height().toInt();
+      final int originalWidth = skImage.width();
+      final int originalHeight = skImage.height();
       box = SkiaObjectBox<CkImage, SkImage>.resurrectable(this, skImage, () {
         final SkImage? skImage = canvasKit.MakeImage(
           SkImageInfo(
@@ -277,13 +277,13 @@ class CkImage implements ui.Image, StackTraceDebugger {
   @override
   int get width {
     assert(_debugCheckIsNotDisposed());
-    return skImage.width().toInt();
+    return skImage.width();
   }
 
   @override
   int get height {
     assert(_debugCheckIsNotDisposed());
-    return skImage.height().toInt();
+    return skImage.height();
   }
 
   @override
@@ -328,8 +328,8 @@ class CkImage implements ui.Image, StackTraceDebugger {
         alphaType: alphaType,
         colorType: colorType,
         colorSpace: colorSpace,
-        width: skImage.width().toInt(),
-        height: skImage.height().toInt(),
+        width: skImage.width(),
+        height: skImage.height(),
       );
       bytes = skImage.readPixels(0, 0, imageInfo);
     } else {
